requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.1.1
pymongo==4.3.3
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
jinja2==3.1.2
python-dotenv==1.0.0
fastapi-cache2[redis]
redis